Store script location
All the file references in a script are relative to
the script file itself, so it makes sense to store
its location (relative to CWD) in the script itself.

Provided we can find the script, we can then track
down all the files it references (always assuming
that the paths are accurate, of course).

darcs-hash:20081222203757-1b04a-fd0eda6bc0d3092e2463e97fc1d59310ff9bbaaf