Transform module to bind together steps
Ignore-this: 8f74a4f24f92463704c9c717f1ea8940
Various steps in the pipeline are hidden behind one transformation from
  Scene -> IO (Panel [Speech])
This process is repeated for each scene in the script. The picture writing
doesn't happen until the next part because it depends on the script as a whole
and other data not stored on a per-scene basis.

darcs-hash:20101027002541-1b04a-444aaaa2a1257305f6619ccd4c6f9c2aedd07ddd