Rewrite script parser in more applicative style
In the process change the format of the preamble
for the scripts, inspired by the Jekyll blogging
style. Maybe I should move away from what I have
right now to something which isn't trying so hard
to be a play or movie script?

Anyway, now it's like

---
title: Romeo and Juliet
author: W Shakespear
---